Address review: qualify std::optional and add explicit int casts in tests
- Qualify optional<std::string> as std::optional<std::string> in matmul_4bits_test.cc (no using namespace std there).
- Add explicit static_cast<int> for K/N passed to MlasBlockwiseQuantizedShape and MlasBlockwiseQuantizedBufferSizes to avoid int64->int narrowing warnings-as-errors.

+++ onnxruntime/test/optimizer/graph_transform_test.cc
@@ -10913,12 +10913,12 @@ TEST_F(GraphTransformationTests, MatMulNBitsBiasFusionDoesNotFuseUnsupportedCuda
 
     int q_rows, q_cols;
     MlasBlockwiseQuantizedShape<float, qbits>(block_size, /* columnwise */ true,
-                                              K, N,
+                                              static_cast<int>(K), static_cast<int>(N),
                                               q_rows, q_cols);
 
     size_t q_data_size_in_bytes, q_scale_size, q_zp_size_in_bytes;
     MlasBlockwiseQuantizedBufferSizes<qbits>(block_size, /* columnwise */ true,
-                                             K, N,
+                                             static_cast<int>(K), static_cast<int>(N),
                                              q_data_size_in_bytes, q_scale_size, &q_zp_size_in_bytes);
 
     auto* A = builder.MakeInput<float>(std::vector{M, K}, "A");
@@ -10973,12 +10973,12 @@ TEST_F(GraphTransformationTests, MatMulNBitsBiasFusionDoesNotFuseCudaRouterShape
 
     int q_rows, q_cols;
     MlasBlockwiseQuantizedShape<float, qbits>(block_size, /* columnwise */ true,
-                                              K, N,
+                                              static_cast<int>(K), static_cast<int>(N),
                                               q_rows, q_cols);
 
     size_t q_data_size_in_bytes, q_scale_size, q_zp_size_in_bytes;
     MlasBlockwiseQuantizedBufferSizes<qbits>(block_size, /* columnwise */ true,
-                                             K, N,
+                                             static_cast<int>(K), static_cast<int>(N),
                                              q_data_size_in_bytes, q_scale_size, &q_zp_size_in_bytes);
 
     auto* A = builder.MakeInput<float>(std::vector{M, K}, "A");
@@ -11025,12 +11025,12 @@ TEST_F(GraphTransformationTests, MatMulNBitsBiasFusionCanDisableCudaRouterFusion
 
     int q_rows, q_cols;
     MlasBlockwiseQuantizedShape<float, qbits>(block_size, /* columnwise */ true,
-                                              K, N,
+                                              static_cast<int>(K), static_cast<int>(N),
                                               q_rows, q_cols);
 
     size_t q_data_size_in_bytes, q_scale_size, q_zp_size_in_bytes;
     MlasBlockwiseQuantizedBufferSizes<qbits>(block_size, /* columnwise */ true,
-                                             K, N,
+                                             static_cast<int>(K), static_cast<int>(N),
                                              q_data_size_in_bytes, q_scale_size, &q_zp_size_in_bytes);
 
     auto* A = builder.MakeInput<float>(std::vector{M, K}, "A");
